chore: Abstracts popup centering code
