+++ client/components/map-controls/NewEventPopup.tsx
@@ -3,6 +3,7 @@ import { LatLng, Popup as LeafletPopup } from "leaflet";
 import { Popup, useMap, useMapEvents } from "react-leaflet";
 import EventForm from "../EventForm";
 import { useSupabase } from "../../context/supabase-provider";
+import { CenterPopup } from "../../app/map/[mapId]/utils";
 
 interface NewEventPopupProps {
 	mapId: string;
@@ -15,19 +16,15 @@ export default function NewEventPopup(props: NewEventPopupProps) {
 	const popupRef = useRef<LeafletPopup | null>(null);
 
 	useMapEvents({
+		// opens a new event popup at the clicked location
 		click(event) {
 			setPosition(event.latlng);
 
-			// wait 20 ms for the popup container to populate in the DOM
+			// wait for the ref to update
 			setTimeout(() => {
-				const popupHeight = popupRef.current?.getElement()?.clientHeight;
-				if (popupHeight) {
-					// convert our lat/lng to pixel values
-					const px = map.project(event.latlng);
-					// translate the y-value by half of the popup's height + the marker height
-					px.y -= popupHeight / 2;
-					// convert back to a lat/lng and fly there, centering the popup in view
-					map.flyTo(map.unproject(px));
+				const popup = popupRef.current;
+				if (popup) {
+					CenterPopup(popup, map, false);
 				}
 			}, 20);
 		},
